requirements.txt
black==23.9.1
boto3==1.28.69
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
isort==5.12.0
pandas==2.1.0
pathspec==0.11.2
psycopg2==2.9.7
requests==2.31.0
requests-oauthlib==1.3.1
sqlalchemy==2.0.22